Chronicle/api/requirements.txt
flask>=2.3
flask-cors>=4.0
flask-jwt-extended>=4.6
pymongo>=4.7
python-dotenv>=1.0
Werkzeug>=3.0

